some editing
